requirements.txt
urllib3==1.25.9
selenium==3.141.0
pillow==7.1.2
pandas==1.2.2
openpyxl==3.0.6
tabula-py==2.2.0
virtualenv==20.4.2